+++ src/session_trace/__init__.py
@@ -3,6 +3,7 @@
 from session_trace.assert_tools import (
     assert_no_tool,
     assert_tool_called,
+    assert_tool_input_contains,
     assert_tool_order,
     assert_write_path,
 )
@@ -12,6 +13,7 @@ __all__ = [
     "ToolCall",
     "assert_no_tool",
     "assert_tool_called",
+    "assert_tool_input_contains",
     "assert_tool_order",
     "assert_write_path",
 ]